aplicacion con ejb y ear

+++ gestion-docente-spring/src/main/java/com/ipartek/formacion/dbms/dao/ClienteDAOImp.java
@@ -11,7 +11,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcCall;
@@ -141,7 +140,7 @@ public class ClienteDAOImp implements ClienteDAO {
 		final String SQL = "CALL clienteInforme(?);";
 		Cliente cliente = null;
 		try{
-			Map<Long, Cliente> clientes = jdbctemplate.query(SQL, new ClienteExtractor(), new Object[] { codigo });
+			Map<Integer, Cliente> clientes = jdbctemplate.query(SQL, new ClienteExtractor(), new Object[] { codigo });
 			cliente = clientes.get(codigo);
 			LOGGER.info("cliente: " +toString());
